update: rename internal modules
