+++ covsirphy/downloading/_db_owid.py
@@ -3,7 +3,7 @@
 
 import pandas as pd
 from covsirphy.util.term import Term
-from covsirphy.downloading.db import _DataBase
+from covsirphy.downloading._db import _DataBase
 
 
 class _OWID(_DataBase):
